requirements.txt
numpy
scikit-learn
scipy
sentence-transformers
pypdf
python-dotenv
rich
typer
rouge-score
nltk
pyyaml
transformers
